% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff.mte.R
\name{eff.mte}
\alias{eff.mte}
\title{Efficient Semiparametric Estimation of Marginal Treatment Effects}
\usage{
eff.mte(A, X, Y, Z, ps, S = 1, alpha = 0.05, interaction = FALSE, trim = 0)
}
\arguments{
\item{A}{a vector of a binary treatment variable taking values 0 and 1}

\item{X}{a vector (if there is only a single covariate) or a matrix (if multiple covariates) of covariate(s), where the k-th column gives the k-th covariate}

\item{Y}{a vector of an outcome variable}

\item{Z}{a vector of an instrumental variable}

\item{ps}{a vector of propensity scores obtained from fitting either a parametric or nonparametric regression of A on (X,Z)}

\item{S}{(optional) a positive integer indicating the order of polynomial of MTE model (default is S=1, i.e. a linear model)}

\item{alpha}{(optional) (1-alpha)*100\% is the nominal coverage of confidence intervals reported in the table of results; default is 0.05}

\item{interaction}{(optional) a logical; if \code{TRUE} then an MTE model allowing for X and V interaction effects is estimated. The default is \code{FALSE}}

\item{trim}{(optional) a non-negative constant between 0 and 1 indicating whether observations with extreme propensity scores less than \code{1-trim} and greater than \code{trim} should be excluded. The default is \code{trim=0} so that no observations are excluded}

\item{pc.pval}{(optional) p-value threshold to select PC-based instruments for each exposure. Note that a very strict p-value threshold may leave very few instruments (perhaps none), so the choice should be made with the sample size in mind. The default choice is 5e-8}

\item{clump.thres}{(optional) R^2 clumping threshold for variant correlations; if unspecified the default value is R^2 = 0.95}

\item{exposure.names}{(optional) a J x 1 character vector of the names of exposures included in the model}
}
\value{
Output is a list containing:
\itemize{
\item \code{gamma_lr} \cr
a vector of efficient MTE parameter \code{gamma} estimates
\item \code{gamma_ols} \cr
a vector of conventional (non-robust) MTE parameter \code{gamma} estimates by a simple linear regression
\item \code{se_gamma_lr} \cr
a vector of standard errors relating to \code{gamma_lr}
\item \code{var_gamma} \cr
an estimate of the asymptotic variance-covariance matrix relating to \code{gamma_lr}
\item \code{se_gamma_ols} \cr
a vector of standard errors relating to \code{gamma_ols}
\item \code{ate_eff} \cr
an efficient estimate of the average treatment effect estimand
\item \code{ate_est} \cr
a conventional (non-efficient, non-robust) estimate of the average treatment effect estimand
\item \code{ate_se} \cr
a standard error relating to \code{ate_eff}
\item \code{att_eff} \cr
an efficient estimate of the average treatment effect on the treated estimand
\item \code{att_est} \cr
a conventional (non-efficient, non-robust) estimate of the average treatment effect on the treated estimand
\item \code{att_se} \cr
a standard error relating to \code{att_eff}
\item \code{atu_eff} \cr
an efficient estimate of the average treatment effect on the untreated estimand
\item \code{atu_est} \cr
a conventional (non-efficient, non-robust) estimate of the average treatment effect on the untreated estimand
\item \code{atu_se} \cr
a standard error relating to \code{atu_eff}
\item \code{asg_eff} \cr
an efficient estimate of the average selection on the gain estimand
\item \code{asg_est} \cr
a conventional (non-efficient, non-robust) estimate of the average selection on the gain estimand
\item \code{asg_se} \cr
a standard error relating to \code{asg_eff}
\item \code{plot.mte} \cr
a plot of the efficient estimate of the MTE function averaged over covariates (reported only if there is a single covariate)
}
}
\description{
A method for heterogeneous treatment effects analysis in a semiparametric MTE model
}
\details{
Efficient semiparametric MTE estimation - for details of the method please see the paper "Efficient semiparametric estimation of marginal treatment effects with genetic instrumental variables"
}
\author{
Ashish Patel, Francis J. DiTraglia, and Stephen Burgess
}
